add StringShortenerHelper
