+++ src/main/java/nadja/url_shortener/controller/RestUrlController.java
@@ -5,18 +5,21 @@ import nadja.url_shortener.dto.UrlDto;
 import nadja.url_shortener.entity.Url;
 import nadja.url_shortener.repo.IUrlRepo;
 import nadja.url_shortener.service.ShortenerService;
+import nadja.url_shortener.service.StringShortenerHelper;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDate;
+
 @RestController
 @RequestMapping("/shortenerUrl")
-public class UrlController {
+public class RestUrlController {
     private final ShortenerService shortenerService;
     private final IUrlRepo urlRepo;
     private final String HOST;
 
-    public UrlController(ShortenerService shortenerService, IUrlRepo urlRepo, @Value("${host}") String host) {
+    public RestUrlController(ShortenerService shortenerService, IUrlRepo urlRepo, @Value("${host}") String host) {
         this.shortenerService = shortenerService;
         this.urlRepo = urlRepo;
         this.HOST = host;
@@ -25,10 +28,13 @@ public class UrlController {
     @PostMapping
     public ShortUrlDto convertLongUrl(@RequestBody UrlDto longUrlDto) {
         Url url = new Url();
-
         url.setLongUrl(longUrlDto.getLongUrl());
+        url.setUserID(0);
+        url.setExpiration_date(LocalDate.now().plusDays(3));
+        String shortenedUrl = new StringShortenerHelper().get_shortURL(longUrlDto.getLongUrl());
+
         urlRepo.save(url);
-        String shortenedUrl = shortenerService.get_shortURL(url.getLongUrl());
+        longUrlDto.setId(url.getId());
         return new ShortUrlDto(HOST + "/" + shortenedUrl);
     }
 }
