+++ src/main/java/nadja/url_shortener/service/ShortenerService.java
@@ -1,76 +1,21 @@
 package nadja.url_shortener.service;
 
-import org.springframework.stereotype.Service;
+import nadja.url_shortener.entity.Url;
+import nadja.url_shortener.repo.IUrlRepo;
 
-import java.nio.ByteBuffer;
-import java.util.Base64;
 
-@Service
 public class ShortenerService {
-    public String get_shortURL(String longUrl) {
-        // murmurhash (long) -> byte[] -> base64 -> substring (3);
-        long hash64bit = hash64(longUrl);
-        byte[] arrByte = longToBytes(hash64bit);
-        String base64String = Base64.getUrlEncoder().encodeToString(arrByte);
+    IUrlRepo urlRepo;
 
-        return base64String.substring(3,base64String.length()-1);
+    //Нужно ли проверять что пользователь на этот длинный адрес уже создавал короткий?
+    //ТОГДА нужна проверка при записи в базу данных
+    //пока просто вносим данные в базу без проверки на уникальность
+    public void save(Url newUrl) {
+        urlRepo.save(newUrl);
     }
 
-    protected long hash64(String text) {
-        final byte[] data = text.getBytes();
-        int seed = 0xe17a1465;
-        int length = data.length;
-        final long m = 0xc6a4a7935bd1e995L;
-        final int r = 47;
-
-        long h = (seed & 0xffffffffl) ^ (length * m);
-
-        int length8 = length / 8;
-
-        for (int i = 0; i < length8; i++) {
-            final int i8 = i * 8;
-            long k = ((long) data[i8 + 0] & 0xff) + (((long) data[i8 + 1] & 0xff) << 8)
-                    + (((long) data[i8 + 2] & 0xff) << 16) + (((long) data[i8 + 3] & 0xff) << 24)
-                    + (((long) data[i8 + 4] & 0xff) << 32) + (((long) data[i8 + 5] & 0xff) << 40)
-                    + (((long) data[i8 + 6] & 0xff) << 48) + (((long) data[i8 + 7] & 0xff) << 56);
-
-            k *= m;
-            k ^= k >>> r;
-            k *= m;
-
-            h ^= k;
-            h *= m;
-        }
-
-        switch (length % 8) {
-            case 7:
-                h ^= (long) (data[(length & ~7) + 6] & 0xff) << 48;
-            case 6:
-                h ^= (long) (data[(length & ~7) + 5] & 0xff) << 40;
-            case 5:
-                h ^= (long) (data[(length & ~7) + 4] & 0xff) << 32;
-            case 4:
-                h ^= (long) (data[(length & ~7) + 3] & 0xff) << 24;
-            case 3:
-                h ^= (long) (data[(length & ~7) + 2] & 0xff) << 16;
-            case 2:
-                h ^= (long) (data[(length & ~7) + 1] & 0xff) << 8;
-            case 1:
-                h ^= (long) (data[length & ~7] & 0xff);
-                h *= m;
-        }
-        ;
-
-        h ^= h >>> r;
-        h *= m;
-        h ^= h >>> r;
-
-        return h;
-    }
-
-    private byte[] longToBytes(long x) {
-        ByteBuffer buffer = ByteBuffer.allocate(Long.BYTES);
-        buffer.putLong(x);
-        return buffer.array();
+    public String get_shortURL(String long_Url) {
+        StringShortenerHelper stringShortenerHelper = new StringShortenerHelper();
+        return stringShortenerHelper.get_shortURL(long_Url);
     }
 }
